Prepare to deploy on lambda AWS
