+++ .env.example
@@ -3,10 +3,6 @@
 # External API URL
 API_URL=your-api-url-here
 
-# Redis Configuration
-REDIS_HOST=redis
-REDIS_URL=redis://redis:6379
-
 # API Request Settings
 MAX_RETRIES=3
 TIMEOUT=5.0
