+++ Dockerfile
@@ -1,6 +1,9 @@
 # Use a lightweight Python image
 FROM python:3.10-slim
 
+# Install build dependencies (gcc, make, etc.)
+RUN apt-get update && apt-get install -y build-essential && rm -rf /var/lib/apt/lists/*
+
 # Set the working directory in the container
 WORKDIR /app
