+++ requirements.txt
@@ -1,28 +1,28 @@
+aiohappyeyeballs==2.5.0
+aiohttp==3.11.13
+aiosignal==1.3.2
+alembic==1.15.1
+amqp==5.3.1
 annotated-types==0.7.0
 anyio==4.8.0
+async-timeout==5.0.1
+attrs==25.1.0
+automaton==3.2.0
+autopage==0.5.2
+bcrypt==4.3.0
 black==25.1.0
+cachetools==5.5.2
 certifi==2025.1.31
-# Web Framework
 fastapi==0.115.11
-uvicorn==0.34.0
-
-# HTTP Client
+fastapi-cache2==0.2.1
 httpx==0.28.1
-
-# Data Validation
+pre_commit==4.1.0
 pydantic==2.10.6
-
-# Testing
+pydantic_core==2.27.2
 pytest==8.3.5
 pytest-asyncio==0.25.3
-
-# Linting & Formatting
-black==25.1.0
-pre-commit==4.1.0
-
-# Environment Variables
 python-dotenv==1.0.1
-
-# Cache
-redis[asyncio]==5.0.4
-fastapi-cache2==0.2.1
+slowapi==0.1.9
+uvicorn==0.34.0
+uvloop==0.21.0
+mangum~=0.19.0
